+++ test/utils/compatibility-test.js
@@ -12,6 +12,7 @@ vows.describe(Compatibility)
         assert.isTrue(options.colors.opacity);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isFalse(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);
         assert.isTrue(options.properties.merging);
@@ -45,6 +46,7 @@ vows.describe(Compatibility)
       'gets merged options': function(options) {
         assert.isTrue(options.colors.opacity);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);
@@ -72,6 +74,7 @@ vows.describe(Compatibility)
       'gets template options': function(options) {
         assert.isFalse(options.colors.opacity);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isTrue(options.properties.colors);
         assert.isTrue(options.properties.iePrefixHack);
         assert.isTrue(options.properties.ieSuffixHack);
@@ -98,6 +101,7 @@ vows.describe(Compatibility)
       'gets template options': function(options) {
         assert.isFalse(options.colors.opacity);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isTrue(options.properties.colors);
         assert.isTrue(options.properties.iePrefixHack);
         assert.isTrue(options.properties.ieSuffixHack);
@@ -134,6 +138,7 @@ vows.describe(Compatibility)
       'gets calculated options': function(options) {
         assert.isTrue(options.colors.opacity);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.iePrefixHack);
         assert.isTrue(options.properties.ieSuffixHack);
@@ -161,6 +166,7 @@ vows.describe(Compatibility)
         assert.isTrue(options.colors.opacity);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isTrue(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);
         assert.isTrue(options.properties.merging);
@@ -187,6 +193,7 @@ vows.describe(Compatibility)
         assert.isTrue(options.colors.opacity);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.backgroundSizeMerging);
+        assert.isFalse(options.properties.backgroundClipMerging);
         assert.isTrue(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);
         assert.isTrue(options.properties.merging);